+++ src/Components/swoole/src/Server/WebSocket/Server.php
@@ -84,8 +84,8 @@ class Server extends Base implements ISwooleWebSocketServer
     protected function createServer(): void
     {
         $config = $this->getServerInitConfig();
-        $this->swooleServer = new \Swoole\WebSocket\Server($config['host'], $config['port'], $config['mode'], $config['sockType']);
-        $this->https = $this->wss = \defined('SWOOLE_SSL') && Bit::has($config['sockType'], \SWOOLE_SSL);
+        $this->swooleServer = new \Swoole\WebSocket\Server($config['host'], (int) $config['port'], (int) $config['mode'], (int) $config['sockType']);
+        $this->https = $this->wss = \defined('SWOOLE_SSL') && Bit::has((int) $config['sockType'], \SWOOLE_SSL);
         $this->http2 = $this->config['configs']['open_http2_protocol'] ?? false;
     }
 
@@ -107,9 +107,9 @@ class Server extends Base implements ISwooleWebSocketServer
     {
         return [
             'host'      => $this->config['host'] ?? '0.0.0.0',
-            'port'      => $this->config['port'] ?? 8080,
-            'sockType'  => $this->config['sockType'] ?? \SWOOLE_SOCK_TCP,
-            'mode'      => $this->config['mode'] ?? \SWOOLE_BASE,
+            'port'      => (int) ($this->config['port'] ?? 8080),
+            'sockType'  => (int) ($this->config['sockType'] ?? \SWOOLE_SOCK_TCP),
+            'mode'      => (int) ($this->config['mode'] ?? \SWOOLE_BASE),
         ];
     }
